backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==1.10.14
pyyaml==6.0.1
google-api-python-client==2.125.0
google-auth==2.29.0
supabase==2.3.5
psycopg[binary]==3.1.18
